Tag v1.4.4
Resolves #5

Resolves #6

+++ 5G_Core_Network/full_install.sh
@@ -31,9 +31,10 @@
 # Exit immediately if a command fails
 set -e
 
+APTVARS="NEEDRESTART_MODE=l NEEDRESTART_SUSPEND=1 DEBIAN_FRONTEND=noninteractive"
 if ! command -v realpath &>/dev/null; then
     echo "Package \"coreutils\" not found, installing..."
-    sudo apt-get install -y coreutils
+    sudo $APTVARS apt-get install -y coreutils
 fi
 
 SCRIPT_DIR=$(dirname "$(realpath "$0")")
@@ -42,6 +43,53 @@ cd "$SCRIPT_DIR"
 # Check if the YAML editor is installed, and install it if not
 if ! command -v yq &>/dev/null; then
     sudo "$SCRIPT_DIR/install_scripts/./install_yq.sh"
+fi
+# Check that the correct version of yq is installed
+if ! yq --version 2>/dev/null | grep -q 'https://github\.com/mikefarah/yq'; then
+    echo "ERROR: Detected an incompatible yq installation."
+    echo "Please ensure the Python yq is uninstalled with \"pip uninstall -y yq\", then re-run this script."
+    exit 1
+fi
+
+echo "Parsing options.yaml..."
+# Check if the YAML file exists, if not, set and save default values
+if [ ! -f "options.yaml" ]; then
+    echo "# Upon modification, apply changes with ./generate_configurations.sh." >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# Configure the MCC/MNC and TAC" >>"options.yaml"
+    echo "plmn: 00101" >>"options.yaml"
+    echo "tac: 7" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# Configure the DNN/APN" >>"options.yaml"
+    echo "dnn: nist-dnn" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# Configure the Single Network Slice Selection Assistance Information (S-NSSAI)" >>"options.yaml"
+    echo "sst: 1" >>"options.yaml"
+    echo "sd: FFFFFF" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# If false, AMF will use the default 127.0.0.5, true: it will use the hostname IP" >>"options.yaml"
+    echo "expose_amf_over_hostname: false" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# Toggle whether or not to include the Security Edge Protection Proxies (SEPP1 and SEPP2)" >>"options.yaml"
+    echo "include_sepp: false" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# Configure the ogstun gateway address for UE traffic" >>"options.yaml"
+    echo "ogstun_ipv4: 10.45.0.0/16" >>"options.yaml"
+    echo "ogstun_ipv6: 2001:db8:cafe::/48" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "ogstun2_ipv4: 10.46.0.0/16" >>"options.yaml"
+    echo "ogstun2_ipv6: 2001:db8:babe::/48" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "ogstun3_ipv4: 10.47.0.0/16" >>"options.yaml"
+    echo "ogstun3_ipv6: 2001:db8:face::/48" >>"options.yaml"
+    echo "" >>"options.yaml"
+    echo "# The use of systemctl can be disabled to support installations within Docker. Before changing this value, it is recommended to uninstall the testbed." >>"options.yaml"
+    echo "use_systemctl: true" >>"options.yaml"
+fi
+
+USE_SYSTEMCTL=$(yq eval '.use_systemctl' options.yaml)
+if [[ "$USE_SYSTEMCTL" == "null" || -z "$USE_SYSTEMCTL" ]]; then
+    USE_SYSTEMCTL="true" # Default
 fi
 
 # Check for open5gs-amfd and open5gs-upfd binaries to determine if Open5GS is already installed
@@ -59,39 +107,43 @@ INSTALL_START_TIME=$(date +%s)
 sudo rm -rf logs/
 
 # Prevent the unattended-upgrades service from creating dpkg locks that would error the script
-if systemctl is-active --quiet unattended-upgrades; then
-    sudo systemctl stop unattended-upgrades &>/dev/null && echo "Successfully stopped unattended-upgrades service."
-    sudo systemctl disable unattended-upgrades &>/dev/null && echo "Successfully disabled unattended-upgrades service."
-fi
-if systemctl is-active --quiet apt-daily.timer; then
-    sudo systemctl stop apt-daily.timer &>/dev/null && echo "Successfully stopped apt-daily.timer service."
-    sudo systemctl disable apt-daily.timer &>/dev/null && echo "Successfully disabled apt-daily.timer service."
-fi
-if systemctl is-active --quiet apt-daily-upgrade.timer; then
-    sudo systemctl stop apt-daily-upgrade.timer &>/dev/null && echo "Successfully stopped apt-daily-upgrade.timer service."
-    sudo systemctl disable apt-daily-upgrade.timer &>/dev/null && echo "Successfully disabled apt-daily-upgrade.timer service."
+if [[ "$USE_SYSTEMCTL" == "true" ]]; then
+    if systemctl is-active --quiet unattended-upgrades; then
+        sudo systemctl stop unattended-upgrades &>/dev/null && echo "Successfully stopped unattended-upgrades service."
+        sudo systemctl disable unattended-upgrades &>/dev/null && echo "Successfully disabled unattended-upgrades service."
+    fi
+    if systemctl is-active --quiet apt-daily.timer; then
+        sudo systemctl stop apt-daily.timer &>/dev/null && echo "Successfully stopped apt-daily.timer service."
+        sudo systemctl disable apt-daily.timer &>/dev/null && echo "Successfully disabled apt-daily.timer service."
+    fi
+    if systemctl is-active --quiet apt-daily-upgrade.timer; then
+        sudo systemctl stop apt-daily-upgrade.timer &>/dev/null && echo "Successfully stopped apt-daily-upgrade.timer service."
+        sudo systemctl disable apt-daily-upgrade.timer &>/dev/null && echo "Successfully disabled apt-daily-upgrade.timer service."
+    fi
 fi
 
 if [ ! -d "open5gs" ]; then
     echo "Cloning Open5GS..."
     ./install_scripts/git_clone.sh https://github.com/open5gs/open5gs.git
 fi
+
 cd $SCRIPT_DIR/open5gs
 
 echo
 echo
 echo "Installing Open5GS..."
-export DEBIAN_FRONTEND=noninteractive
 # Modifies the needrestart configuration to suppress interactive prompts
-if [ -f "/etc/needrestart/needrestart.conf" ]; then
-    if ! grep -q "^\$nrconf{restart} = 'a';$" "/etc/needrestart/needrestart.conf"; then
-        sudo sed -i "/\$nrconf{restart} = /c\$nrconf{restart} = 'a';" "/etc/needrestart/needrestart.conf"
-        echo "Modified needrestart configuration to auto-restart services."
-    fi
+if [ -d /etc/needrestart ]; then
+    sudo install -d -m 0755 /etc/needrestart/conf.d
+    sudo tee /etc/needrestart/conf.d/99-no-auto-restart.conf >/dev/null <<'EOF'
+# Disable automatic restarts during apt operations
+$nrconf{restart} = 'l';
+EOF
+    echo "Configured needrestart to list-only (no service restarts)."
 fi
-export NEEDRESTART_SUSPEND=1
 
 sudo "$SCRIPT_DIR/./install_scripts/install_mongodb.sh"
+sudo "$SCRIPT_DIR/./install_scripts/start_mongodb.sh"
 
 # Check and create the open5gs user and group if they don't exist
 if ! getent passwd open5gs >/dev/null; then
@@ -107,11 +159,11 @@ sudo usermod -a -G open5gs open5gs
 echo "Installing dependencies for building Open5GS..."
 
 # Code from (https://open5gs.org/open5gs/docs/guide/02-building-open5gs-from-sources#building-open5gs):
-sudo apt-get install -y python3-pip python3-setuptools python3-wheel ninja-build build-essential flex bison git cmake libsctp-dev libgnutls28-dev libgcrypt-dev libssl-dev libmongoc-dev libbson-dev libyaml-dev libnghttp2-dev libmicrohttpd-dev libcurl4-gnutls-dev libnghttp2-dev libtins-dev libtalloc-dev meson
+sudo $APTVARS apt-get install -y python3-pip python3-setuptools python3-wheel ninja-build build-essential flex bison git cmake libsctp-dev libgnutls28-dev libgcrypt-dev libssl-dev libmongoc-dev libbson-dev libyaml-dev libmicrohttpd-dev libcurl4-gnutls-dev libnghttp2-dev libtins-dev libtalloc-dev meson
 if apt-cache show libidn-dev >/dev/null 2>&1; then
-    sudo apt-get install -y --no-install-recommends libidn-dev
+    sudo $APTVARS apt-get install -y --no-install-recommends libidn-dev
 else
-    sudo apt-get install -y --no-install-recommends libidn11-dev
+    sudo $APTVARS apt-get install -y --no-install-recommends libidn11-dev
 fi
 
 rm -rf build
